+++ scripts/Wall.js
@@ -1,6 +1,5 @@
 /* globals
-canvas,
-game
+
 */
 "use strict";
 /* eslint no-unused-vars: ["error", { "argsIgnorePattern": "^_" }] */
@@ -72,7 +71,7 @@ function updateWall(wallD, changes, _options, _userId) {
  */
 function deleteWall(wallD, _options, _userId) {
   // The document.object is now null; use the id to remove the wall.
-  SCENE_GRAPH.removeWall(document.id);
+  SCENE_GRAPH.removeWall(wallD.id);
   SCENE_GRAPH.updateCyclePolygons();
 }