+++ scripts/regions/AddTerrainRegionBehaviorType.js
@@ -1,5 +1,6 @@
 /* globals
-RegionBehaviorType
+CONFIG,
+foundry,
 */
 /* eslint no-unused-vars: ["error", { "argsIgnorePattern": "^_" }] */
 "use strict";